The implementation of server methods
Taking assumption that remote call should be passed by reference the
client interfaces are also implemented on the server side.
Some further comments regarding methods to generate the document should
be added.
The asynchronized timer ought to be added on the server side to provide
constant refresh of randomly generated position of fly.

+++ src/impl/IGameClient.java
@@ -4,6 +4,6 @@ import java.rmi.Remote;
 import java.rmi.RemoteException;
 
 public interface IGameClient extends Remote{
-	public void recieveFlyHunted(String playerName, int newPoints) throws RemoteException;
+	public void recieveFlyHunted(String playerName, long newPoints) throws RemoteException;
 	public void recieveFlyPosition(int x, int y) throws RemoteException;
 }
